Remove unnecessary package references and content includes
This commit involves cleaning up the QuickBot.csproj file. The dotnet-gitversion package reference was removed due to an upgrade to GitVersion.MsBuild ver.6.0.0-beta.3 which fulfills the same role. Furthermore, the unnecessary references to debug objects and cache files in the ItemGroup field were removed to reduce clutter and improve readability. This is in line with best practices for maintaining a clear and understandable .csproj file.